Updated Admin settings
Fixed UI and implementing user Authorization

+++ Gaant Chart/Models/Model.cs
@@ -8,14 +8,14 @@ namespace Gaant_Chart.Models
 {
     public class Model
     {
-        public int rowid { get; set; }
+        public long rowid { get; set; }
         public String modelName { get; set; }
         public DateTime startDate { get; }
 
         public List<Task> tasks = new List<Task>();
         public int lastCompletedTaskId { get; set; }
 
-        public Model(int rowid, String modelName, DateTime startDate)
+        public Model(long rowid, String modelName, DateTime startDate)
         {
             this.rowid = rowid; 
             this.modelName = modelName;
